Fixed complie issues

+++ firmware/src/services/iot/cloud/mqtt_packetPopulation/mqtt_iotprovisioning_packetPopulate.h
@@ -31,10 +31,10 @@
 
 #include <stdbool.h>
 #include <stdint.h>
-#include "../../config/IoT_Sensor_Node_config.h"
+#include "iot_config/cloud_config.h"
 
 void MQTT_CLIENT_iotprovisioning_publish(uint8_t *data, uint16_t len);
-void MQTT_CLIENT_iotprovisioning_receive(uint8_t *data, uint8_t len);
+void MQTT_CLIENT_iotprovisioning_receive(uint8_t *data, uint16_t len);
 void MQTT_CLIENT_iotprovisioning_connect(char* deviceID);
 bool MQTT_CLIENT_iotprovisioning_subscribe();
 void MQTT_CLIENT_iotprovisioning_connected();
